requirements.txt
python-dotenv
elasticsearch
textblob
tweepy
